% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatVenue.R
\name{formatVenue}
\alias{formatVenue}
\title{COnverts Betfair formatted venues to OddsChecker compatible venues}
\usage{
formatVenue(venue)
}
\arguments{
\item{venue}{string. The name of the horse racing venue to be converted.}
}
\value{
A string of the Oddschecker compatible venue is returned.
}
\description{
\code{formatVenue} is a simple function that converts Betfair API venues to 
OddsChecker compatible venues (lower case hypehnated separated strings). The
venue nomenclature appears quite consistent in the UK and Ireland. However,
especially in the Americas, Betfair and Oddschecker may have different names
for the same venue. To allow Betfair API data to be combined with scraped
OddsChecker data, we need to ensure that venue names match up. The list of
venues is not necessarily exhausitive and potential omissions should be
raised as issue or pull request.
}
\examples{
\dontrun{

# Simple venue conversion

formatVenue("Emerald Downs")

# Example where Betfair and Oddschecker differ significantly

formatVenue("Parx Racing At Philadelphia Park")
}
}

