% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracToDec.R
\name{fracToDec}
\alias{fracToDec}
\title{Convert odds from fraction to decimal}
\usage{
fracToDec(price)
}
\arguments{
\item{price}{String. A string representing the odds of a bet, expressed in
fractional form (e.g. 5/4).}
}
\value{
The output of the function is a numeric. In the event of "SP", 0.0 is
  returned, while "" is conveted -1.0.
}
\description{
\code{partialMatch} converts betting odds from fraction to decimal. By
default, Oddschecker provides odds as fractions, while the exchange utilise
decimal odds. To allow easier comparision, we need to convert the fractional
odds to decimal. The function also returns negative under specific scenarios
(e.g. no odds).
}
\examples{
\dontrun{

# 2.25 is the decimal form of "5/4"

fracToDec("5/4")

# We can map a vector of fractional odds to their decimal form equivalent

sapply(c("5/4","8/11","4/6","5","22/2"),fracToDec)

# Finally, sometimes the odds are not provided, or at least not in fractional form. To account for this, specific numbers are returned.

sapply(c("EVS","SP",""),fracToDec)

}
}

